Introduce the first SPI
It is to be the lock provider SPI. So that retrieving and intializing lock
implementations can be shielded from API impl, so that by changing SPI
API impl should be useable across different layers of an application.

Add dependency to content-api and smart-bean-util.

Next the SPI implementation will be moved to a separate module

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>