Introduce SPI for CMS Implementation
The idea is to abstract implementation of API and implementation of API
used by the API implementation, which we are terming as SPI.

In next step SPI and some SPI implementations will be moved to two separate
modules. API Impl will depend on SPI and SPI Impl will depend of SPI only.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>