Change content definition schema to be more simple and resource oriented
The main conceptual change to the content type perception is that every
workspace will only have content(s). Content will have fields which
individually will have MIME-type, making them standalone resources; again
content itself is a resource; thus we can define content as a resource
formed of several other resources. As a hypermedia, content will have 2
representations by default, complete and minimal, where the minimal will
simply point to the field resources using some form of a link.

As now we have designed content to be and have resources, as a CMS, we need
to support representations of resources as well. So the next step would be
to accommodate representation generation of content in some manner, so that
user can have their own representation in whatever content type desired.

This change will trigger change of API, SPI for content definition and
their implementations.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>