Modify grammer to make reference to content uniform and include workspace
It is needed that contents can associate and extend from contents across
workspaces, provided the workspaces are mutually "friend". Schema should
allow referring to content, either for association or for inheritance, in a
uniform manner and thus complexType contentPointer which is used for parent
and definition (in content for association.

Modify the samples to satisfy the new grammer.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>