Modify the Lock API, SPI to remove dependency from Java Concurrency API
The main reason for not using the concurrency API is that we need lock
support across a distributed platform. That is, a domain can be locked
across any server, i.e. is Content 1 is locked from Server 1 it should not
be lockable by another client from Server 2.

We will later have to have a HBase based lock implementation.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>