Add search field name retrieval from field definition
A Field definition should be able to retrieve the name of the field to be
used for search. Please note that there will 2 SPI functionalities for it,
first a prefix for fields, content type, status, dates and suffix based on
the SearchDef. It will be added to the SPI later.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>