Fix Filter API
Filter API was missing getter methods to retrieve the settings, so add
getters for every setting.

The Filter implementation was all screwed up in setting. E.g., adding and
removing was doing a 'set' rather than adding or removing. Fix that so
that backup data structure is a collection adding adding and removing.

Add a factory method in content loader to create filter instance.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>