Add new SPI for API Impl to implement, a persistent domain SPI
Persistable* refers to SPI which API Impl implements to provide instances
of the API domains with SPI specification.

Persistent* refers to DTOs used to persist directly into persistent storage
and used in SPI implementation only and should not be used outside of it.

Thus change PersistableContentType to PersistentContentType and also add
PersistableContentType SPI. The API Impl provides an implementation of the
new PersistableDomainFactory SPI and is injected by integrator module.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>