Ensure that only representation/variation info can be read
It is needed to ensure that only the representation and variation info
needs to be readable while the others are absent and the converted should
be able to handle it.

For this ensure that if data is missing cell is checked for null. And
organization by name is peformed for data if data is loaded.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>