Dependency inject asynchronous event publication/subscription
To fallback to old way of indexing simply set
com.smartitengineering.cms.event.async=false in the configuration

If cluster of CMS is being set ensure that following is set only in one of
the nodes.
com.smartitengineering.cms.event.async.subscribe=true
Preferable that node should not be used for reading or writing.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>