(De)Serialize composite field values to JSON
Basically the JSON representation will be persisted against the composite field in
HBase. For this add getJsonNode to collection field which returns an object to
leave it to the implementation to specify what type of object or what library will
be used. In this case use ArrayNode and ObjectNode for collection and composite
fields respectively.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>