Add test case creating content with composite fields
This chance encompasses fix regarding parsing and forming JSON data for
both the web service and persistence layer. The primary bug is not taking
into account special cases such as Collection of Composites or Composite
of Collection.

In addition fix the indexed field name for composite fields having '.' as the
separator for field hierarchy.

Test both content creation via HTML Form and JSON. The next step would
be to implement nested validation for composite field and test search in
solr.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>