Add content definition type to content type API to state the definition type
Currently there are 4 types of definition supported. They are abstract type &
component, concrete type & component. The pre-conditions in application
re that child of type can not be a component and vice versa. In addition
if a child does not mention its definition it inherits the concrete def type of
its parent's def type.

Def type is indexed as status of content type. Content can be created of
only concrete type's. But there is no restriction on relationship, though
it is immaterial of have component in a has-a relation there is no
validation check for that in the content the definition. Similarly there is
no constraint on the content type definition within a composite field. 
These checkings are next step.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>