Generate the DTO conversion helper and private google guice module
This is the first step in generating the DAO imlementation for DTOs. For
this change abstract repo adapter helper to relax the generic restriction
on abstract repo domain and do the same for dao impl as well.

Next step is to generate the google guice bindings and then the helper
code. The helper code primarily concerns generating Field and FieldValue
for the content.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>