Generate the Guice configuration for binding DAOs and exposing the APIs
CommonDao, CommonReadDao and CommonWriteDao are exposed while
the adapter and workspace id are hold back. The module is a private one to
shield unexposed beans.

Make workspace id mandatory and extend its role to be the container to
which contents are saved, that is it is DId to adapter and DAO Impl.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>