Generate code to convert from content to bean
The code is raw and uses a lot of copy'n'paste, but works :). Will need to
develop better alternates once the whole demo works as expected.

Changed the content type definition in demo to add more scenarios and
fixed the link to XSD of content type.

Next step is to convert from bean to content.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>