Implement related domain object to content id conversion
For this added a helper method as getContentId(String, String) in the
abstract helper, its main purpose is to create content id from given
workspace id and content id string. If workspace id is null then it will use
the default workspace id.

Next step is to convert composite objects.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>