Generate master module for aggregating all modules into one
This is primarily done to simplify initialization. Now the client code
depending on this module can simply use the MasterModule which will
be present under all circumstances.

Also add a test case to ensure that the injection works as expected.
Next step is to verify that beans can actually be saved as perceived.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>