Get an integrated test environment running
As an immediate benefit we fixed a couple of bugs in the code -
1. Composite field value not being set in setField method
2. When ID is generated it is not being set back to the bean

Add all the dependencies to get the CMS started. Basically the webservice
is not required just to get things started by using what we had we installed
it in the test case, but in general it is not required when CMS is being used
as a content repository.

Next we need more test cases to be sure that it works as expected. In
addition we will also need to implement query based read methods.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>