Fix setting of simple type into bean
Due to a bug in code generator instead of checking val != null it was
checking val == null, thus simple values not being set. This bug is fixed
and thus writing and reading of beans is working smooth!

A more complicated bean is saved properly in a test case and we should
done with it for now, but we will also need to add more read methods in
common read DAO

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>