Move all logic for field name generation to the SPI implementation
This is a huge change and is accumulation of several changes. So first
the inspiration for the change -
  Generate solr search field name from content type field definition. This
  will enable us to provide a Criteria API of Hibernate like feature to user
  to query to CMS, but it is to be noted with due care that if the field is
  not indexed then it will duely be ignored.

The first change is to introduce the concept of parentContainer in
FieldDef API which enables a field def to keep track from where it came to
being. This also led to adding a new method to ContentTypeLoader API
which creates mutable field def with a parent container field def.

The second change is to set parent container when parsing XML in XML
Parser and in content type object converter (for when reading content 
types).

The third change is to move all field name generation code to SPI for
field name generation and thus introduce a new method to generate
field name without the type specific solr dynamic field suffix. As a result
prefix in content helper for solr document conversion is used for only
related content's fields.

The fourth change is rather a bug fix in POM configuration where event
hub was not being shutdown once the test finishes, the fix is 2 steps -
  1. Change the port number of event hub in code generation demo
  2. Add POM plugin execution for stopping jetty when test finishes.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>