Add POJO member to content field query conversion
It basically converts from POJO like query API to SOLR fields. It uses the CMS
API to generate field names and uses nested query parameters for complex
and related objects.

Generate constants in POJOs so that field names need to be hand-coded,
rather be used as constants. Add API to content loader to escape chars
and support order by.

The test case is insufficient but at least validates that it works in most basic
form. We will add more test case as bugs get found.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>