Add support for passing field value to JasperReport representations
By default the POJO passed to JasperReports is a wrapper collection over the
content for which representation is request, i.e. size of the collection is 1.
Instead it might be more fruitful to pass a field to the report, e.g., a
collection of composite field. It is to be noted that the field could be a
collection or non-collection field, in case of non-collection field it will be
wrapped in a collection and in case of collection field it will be passed as it
is.

The param key 'pojo_field' will specify a field name which would be passed to
the representation instead of the content. It is to be noted that field value
will be passed to the report as bean data source instead of the field.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>