Fix index field names for composite field's embedded content type fields
The bug was due to the fact that field definitions from embedded content
type do not have their parent container field definition set, thus when name
is being generated the prefixes are missing.

Change the CompositeDataType API to introduce embedded content
data type which encompasses both the field definition it is in and the data
type itself. So when retrieving its fields the parent container is set in them
by cloning them. Also add factory method for them so that the creation
is not direct for external users.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>