Refactor SPI modules to make implementations independent
The idea of this refactor is to be able to swap small implementations
in the whole SPI implementation, for example only switch HBase impl
with RDBMS impl or Cassandra impl that should be sufficient the rest
of implementations will simply fall into places.

The next step is to refactor the solr related codes which seem to have
some hard bindings with the default SPI implementation.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>