Introduce a new SPI Impl only API to decouple search and persistence impl
The idea is that the default SPI implementations should be usable without
each other, that is teh search code should be usable without the HBase
implementation being needed. Thus abstracted dependency through the
new SearchBeanLoader API which is part of the SPI implementation.

Next step would be to move the search related code to another module
completing the decoupling of modules.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>