Fix properly reading associated contents as objects
Now whenever there is any association it will be represented by a map
of DAOs, mapping from content type id to DAO. So in case of inheritence
it will have all concrete types of the hierarchy in the map during DI code
generation; in case of single concrete type, i.e. inexistence of inheritence
the map will contain a single entry. During unmarshalling the content its
content type id will determine which DAO will be used to unmarshall the
associated content.

Alternative to map we could have chosen if-elseif-else block for the
unmarshalling of association, but that would be in runtime of the DAO,
map will give a constant time for lookup hence better performance.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>