Add Enumeration data type support to API
There is a now a special data type representing enumeration and it
takes choices of which data is to be allowed. Field value for enum
will be of type StringFieldValue. There is validation added for enum
fields in API validation to ensure that a value represents exactly one
of the choices presented in a case sensitive manner.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>