Add support to configure content co processors in content type API/XML
Now content co-processors can be configured through the XML config doc
through web service or API and through the native Java API. There is test
case for parsing and persistence of content type with this configuration.

One important thing to note is that current implementations have 2 things
to note:
1. Execution will be determined through priority of processor; when configd
    through XML its determined by the order of the specification
2. Content co processor is the only resource definition not having a mime
    type as it does have any output for the client, thus effectively making
    mime type optional.

Next step is to invoke co processors in their respective phases in proper
order.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>