Add Sequence API with full API/SPI implementations
Sequence works with a simple philosophy of incrementing or
decrementing long. Where user decides the next delta for the sequence.
This means the client of sequence API has to be responsible enough to
make sure it specifies delta sufficient to server its purpose. Also to be
noted is that sequence is thread-safe in and across JVMs

Next steps are -
 * Prepare DI into workspace service of SPI for sequence
 * Create web service resources for sequence
 * Write web service clients utilizing sequence
 * Write test case verifying sequence works in multi-thread env

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>