Index sequence so that they can be searched
Since we decided to save sequence into a separate table for performance
related issue, it is necessary to find sequences for a workspace and that
we are doing using a Solr search. The earlier change we performed made
sure that events were getting fired for sequence changes and now with
this change we ensure that sequence gets indexed.

In process we needed to introduce sequence id in the API and change earlier
implementation of service accordingly.

Next we will test sequence and its behavior via the API before moving to
web service.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>