Skip update event for sequence
The primary reason is when sequences will be used they will be used
a lot and that too simply change the value. It does not make sense
to record each change as it will be become invalid by the time it reaches
the consumer of the event. Thus for now its skipped knowingly. Not to
mention that it does not harm to save processing thousands of events.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>