Avoid using 2 jetty instances for test case
The primary goal for this is to be able to run both CMS and Solr from a
single container and in order to reduce the resource footprint.

Though we have managed to successfully integrate into one container,
this has come at a price of having to delete some JARs' manually through
maven exec plugin from the hub webapp. This has added 2 minus'
 1. The command is *nix dependent
 2. The versions could be different
The reason we yet chose to go ahead is *nix is our targeted platform for
CMS and versions of EventHub and CMSs' Jersey, Hadoop and HBase are
not supposed to be different, in fact, they should not be different.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>