Fix parsing collection of composite fields' name
The bug was due to having collection of composite within another collection
of composite field. This made it possible to have ambiguity over the default
fields, which are supposed to be ignored, as sent cms-editor. Thus made it
stricter and crystal clear that demarcation of composite field collection
would ONLY be an integer.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>