Add new plugin to manage all CMS tools
This will basically work as the development platform to start and stop
all tools and suites required to work with CMS. Currently the start
mojo is imlemented. Next we will need to implement the stop mojo.

Add sample configuration for the new plugin in the code generator demo
to act as the base for sample usage.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>