Implement 'stop' mojo to make it possible to embed it in Maven POM
The 'stop' mojo will also help comlete the usage of the plugin as a dev
container. It simply sends the 'stop' signal and waits for the response
which confirms that the stop has been completed.

Also configure this plugin in demo POM. The result was deletion of various
plugins responsible for copying and extracting solr conf and wars. Also
reduced drastically is dependencies and codes in test case start up and
tear down.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>