Make jetty port unconfigurable to avoid configuration issues for smart cms
Since CMS depdends on port configuration in form of URI for CMS and Hub
it would be foolish to allow jetty port configurable while the properties file for
the configuration ain't. We need to be able to create the properties file in
runtime so that it does not conflict and for now thats not necessary.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>