Introduce local lock registrar
Lock on a Key will have to be in 2 steps
1. Lock across JVM
2. Lock across Cluster
It needs to be in this sequence exactly so that there is no race
condition in acquiring local lock. Local lock will be released after the
configured timeout period. The timeout period in local lock registrar
should be greater than that of local lock handler.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>