Fix bug in generated helper code for collection of composite and content
The bug with collection of composite was in converting from bean to content,
the issue was the individual composite field value was not being added to
the collection field value.

The bug with collection of content was in the reading process. The lookup for
read dao is performed using the content type's toString instead of its id's
toString. So the additional invoke is added and thus now it reads content
properly as well.

Add test case to verify these bugs and also ensure that collection of primitive
types are also persisted and read properly

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>