Fix velocity template based representations not being generated
The problem was when a representation template is accessed more
than once, which is the case due to generators being cached, for
velocity it gets invoked with a consumed template reader after the
first invocation, thus caused blank template to be assumed by velocity.

This commit fixes that issue by ensuring template reader are local and
are initiated on individual invocations.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>