Fix search on dates through web service
There were 2 issues. First, if no restriction other than date is applied then
no result is return due to bug in the query generated. Now a dummy
query is added in front of so nullify the bug. Second shorter comparator
first checking caused number format exception so the order of checking
for comparator is fixed.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>