Enhance configuration for implicit lock await and cache disabling
There was no configuration to specify that implicit locks should wait for
attaining it or exit instantaneously; thus for multithreaded scenario of
editing a bean was an impossible scenario with implicit locking. With the
new configuration on deployment it can be configured; by default it will
wait for a lock in case of implicit locking.

Also add configuration to disable caching; under certain circumstance it
might be useful; e.g. if we know that the op will write intensive from a
particular app. Also for debugging cache related bugs it can be handy.

Signed-off-by: Imran M Yousuf <imyousuf@smartitengineering.com>