Add implementation for full Smart CMS API initialization
This involves adding dependency to Smart CMS DI, Smart CMS HBase table creation,
the workspace creation (if not exists) and content type update. Also to be noted is if the
initialization is done multiple times it will only try to create the workspace and content
types but not the HBase tables or Smart CMS API initialization.

The addition of dependency ensures that no additional dependency is needed for
working with Smart CMS except for the Content Repositroy DAO Impl dependency.

Signed-off-by: Imran M Yousuf <imyousuf@smitsol.com>
Issue #SMARTCMS-2 - Make it easier for creating workspace and content type during startup