Implement transaction manager
In process implement transaction and also enhance its API to allow transaction completion
observer. It is required for notifying the manager of a transaction completion. In process of
this implementation TransactionService SPI which will integration the transaction API with
the Common DAO API.

Also enhance the cache memory SPI to allow lookup of all transaction memories related to
a particular transaction.

TransactionImpl still lacks test case and it will be the next thing to do. After that we will
implement the transaction service commit, rollback and dao integration.

Signed-off-by: Imran M Yousuf <imyousuf@smitsol.com>
Issue #SMARTCMS-50 - Allow In-Memory transaction capabilities for Smart CMS DTOs